power-node/requirements.txt
pi-ina219>=1.3.0
paho-mqtt>=1.6.1
pyyaml>=6.0
